+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/content_blocker/ContentBlockerHandler.java
@@ -2,31 +2,33 @@ package com.pichillilorenzo.flutter_inappwebview.content_blocker;
 
 import android.os.Build;
 import android.os.Handler;
+import android.text.TextUtils;
 import android.util.Log;
 import android.webkit.WebResourceResponse;
 
 import androidx.annotation.Nullable;
 
-import com.pichillilorenzo.flutter_inappwebview.webview.in_app_webview.InAppWebView;
-import com.pichillilorenzo.flutter_inappwebview.plugin_scripts_js.JavaScriptBridgeJS;
 import com.pichillilorenzo.flutter_inappwebview.Util;
+import com.pichillilorenzo.flutter_inappwebview.plugin_scripts_js.JavaScriptBridgeJS;
+import com.pichillilorenzo.flutter_inappwebview.types.WebResourceRequestExt;
+import com.pichillilorenzo.flutter_inappwebview.webview.in_app_webview.InAppWebView;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
+import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.regex.Matcher;
 
 import javax.net.ssl.SSLHandshakeException;
-
-import okhttp3.Request;
-import okhttp3.Response;
 
 public class ContentBlockerHandler {
     protected static final String LOG_TAG = "ContentBlockerHandler";
@@ -48,9 +50,13 @@ public class ContentBlockerHandler {
     }
 
     @Nullable
-    public WebResourceResponse checkUrl(final InAppWebView webView, String url, ContentBlockerTriggerResourceType responseResourceType) throws URISyntaxException, InterruptedException, MalformedURLException {
+    public WebResourceResponse checkUrl(final InAppWebView webView, WebResourceRequestExt request,
+                                        ContentBlockerTriggerResourceType responseResourceType)
+            throws URISyntaxException, InterruptedException, MalformedURLException {
         if (webView.customSettings.contentBlockers == null)
             return null;
+
+        String url = request.getUrl();
 
         URI u;
         try {
@@ -182,36 +188,85 @@ public class ContentBlockerHandler {
                         if (scheme.equals("http") && (port == -1 || port == 80)) {
                             String urlHttps = url.replace("http://", "https://");
 
-                            Request mRequest = new Request.Builder().url(urlHttps).build();
-                            Response response = null;
+                            HttpURLConnection urlConnection = Util.makeHttpRequest(urlHttps, request.getMethod(), request.getHeaders());
+                            if (urlConnection != null) {
+                                try {
+                                    byte[] dataBytes = Util.readAllBytes(urlConnection.getInputStream());
+                                    if (dataBytes == null) {
+                                        return null;
+                                    }
+                                    InputStream dataStream = new ByteArrayInputStream(dataBytes);
 
-                            try {
-                                response = Util.getBasicOkHttpClient().newCall(mRequest).execute();
-                                byte[] dataBytes = response.body().bytes();
-                                InputStream dataStream = new ByteArrayInputStream(dataBytes);
+                                    String encoding = urlConnection.getContentEncoding();
+                                    String contentType = urlConnection.getContentType();
+                                    if (contentType == null) {
+                                        contentType = "text/plain";
+                                    } else {
+                                        String[] contentTypeSplitted = contentType.split(";");
+                                        contentType = contentTypeSplitted[0].trim();
+                                        if (encoding == null) {
+                                            encoding = (contentTypeSplitted.length > 1 && contentTypeSplitted[1].contains("charset="))
+                                                    ? contentTypeSplitted[1].replace("charset=", "").trim()
+                                                    : "utf-8";
+                                        }
+                                    }
 
-                                String[] contentTypeSplitted = response.header("content-type", "text/plain").split(";");
-
-                                String contentType = contentTypeSplitted[0].trim();
-                                String encoding = (contentTypeSplitted.length > 1 && contentTypeSplitted[1].contains("charset="))
-                                        ? contentTypeSplitted[1].replace("charset=", "").trim()
-                                        : "utf-8";
-
-                                response.body().close();
-                                response.close();
-
-                                return new WebResourceResponse(contentType, encoding, dataStream);
-
-                            } catch (Exception e) {
-                                if (response != null) {
-                                    response.body().close();
-                                    response.close();
-                                }
-                                if (!(e instanceof SSLHandshakeException)) {
-                                    e.printStackTrace();
-                                    Log.e(LOG_TAG, e.getMessage());
+                                    String reasonPhrase = urlConnection.getResponseMessage();
+                                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && reasonPhrase != null) {
+                                        Map<String, String> responseHeaders = new HashMap<>();
+                                        for (Map.Entry<String, List<String>> responseHeader : urlConnection.getHeaderFields().entrySet()) {
+                                            responseHeaders.put(responseHeader.getKey(), TextUtils.join(",", responseHeader.getValue()));
+                                        }
+                                        return new WebResourceResponse(contentType,
+                                                encoding,
+                                                urlConnection.getResponseCode(),
+                                                reasonPhrase,
+                                                responseHeaders,
+                                                dataStream);
+                                    } else {
+                                        return new WebResourceResponse(contentType,
+                                                encoding,
+                                                dataStream);
+                                    }
+                                } catch (Exception e) {
+                                    if (!(e instanceof SSLHandshakeException)) {
+                                        e.printStackTrace();
+                                    }
+                                } finally {
+                                    urlConnection.disconnect();
                                 }
                             }
+
+//                            Request mRequest = new Request.Builder().url(urlHttps).build();
+//                            Response response = null;
+//
+//                            try {
+//                                response = Util.getBasicOkHttpClient().newCall(mRequest).execute();
+//                                byte[] dataBytes = response.body().bytes();
+//                                InputStream dataStream = new ByteArrayInputStream(dataBytes);
+//
+//                                String[] contentTypeSplitted = response.header("content-type", "text/plain").split(";");
+//
+//                                String contentType = contentTypeSplitted[0].trim();
+//                                String encoding = (contentTypeSplitted.length > 1 && contentTypeSplitted[1].contains("charset="))
+//                                        ? contentTypeSplitted[1].replace("charset=", "").trim()
+//                                        : "utf-8";
+//
+//                                response.body().close();
+//                                response.close();
+//
+//                                return new WebResourceResponse(contentType, encoding, dataStream);
+//
+//                            } catch (Exception e) {
+//                                if (response != null) {
+//                                    response.body().close();
+//                                    response.close();
+//                                }
+//                                if (!(e instanceof SSLHandshakeException)) {
+//                                    e.printStackTrace();
+//                                    Log.e(LOG_TAG, e.getMessage());
+//                                }
+//                            }
                         }
                         break;
                 }
@@ -221,48 +276,36 @@ public class ContentBlockerHandler {
     }
     
     @Nullable
-    public WebResourceResponse checkUrl(final InAppWebView webView, String url) throws URISyntaxException, InterruptedException, MalformedURLException {
-        ContentBlockerTriggerResourceType responseResourceType = getResourceTypeFromUrl(url);
-        return checkUrl(webView, url, responseResourceType);
+    public WebResourceResponse checkUrl(final InAppWebView webView, WebResourceRequestExt request) throws URISyntaxException, InterruptedException, MalformedURLException {
+        ContentBlockerTriggerResourceType responseResourceType = getResourceTypeFromUrl(request);
+        return checkUrl(webView, request, responseResourceType);
     }
 
     @Nullable
-    public WebResourceResponse checkUrl(final InAppWebView webView, String url, String contentType) throws URISyntaxException, InterruptedException, MalformedURLException {
+    public WebResourceResponse checkUrl(final InAppWebView webView, WebResourceRequestExt request, String contentType) throws URISyntaxException, InterruptedException, MalformedURLException {
         ContentBlockerTriggerResourceType responseResourceType = getResourceTypeFromContentType(contentType);
-        return checkUrl(webView, url, responseResourceType);
+        return checkUrl(webView, request, responseResourceType);
     }
 
-    public ContentBlockerTriggerResourceType getResourceTypeFromUrl(String url) {
+    public ContentBlockerTriggerResourceType getResourceTypeFromUrl(WebResourceRequestExt request) {
         ContentBlockerTriggerResourceType responseResourceType = ContentBlockerTriggerResourceType.RAW;
+        String url = request.getUrl();
 
         if (url.startsWith("http://") || url.startsWith("https://")) {
             // make an HTTP "HEAD" request to the server for that URL. This will not return the full content of the URL.
-            Request mRequest = new Request.Builder().url(url).head().build();
-            Response response = null;
-            try {
-                response = Util.getBasicOkHttpClient().newCall(mRequest).execute();
-
-                if (response.header("content-type") != null) {
-                    String[] contentTypeSplitted = response.header("content-type").split(";");
-
-                    String contentType = contentTypeSplitted[0].trim();
-                    String encoding = (contentTypeSplitted.length > 1 && contentTypeSplitted[1].contains("charset="))
-                            ? contentTypeSplitted[1].replace("charset=", "").trim()
-                            : "utf-8";
-
-                    response.body().close();
-                    response.close();
-                    responseResourceType = getResourceTypeFromContentType(contentType);
-                }
-
-            } catch (Exception e) {
-                if (response != null) {
-                    response.body().close();
-                    response.close();
-                }
-                if (!(e instanceof SSLHandshakeException)) {
+            HttpURLConnection urlConnection = Util.makeHttpRequest(url, "HEAD", request.getHeaders());
+            if (urlConnection != null) {
+                try {
+                    String contentType = urlConnection.getContentType();
+                    if (contentType != null) {
+                        String[] contentTypeSplitted = contentType.split(";");
+                        contentType = contentTypeSplitted[0].trim();
+                        responseResourceType = getResourceTypeFromContentType(contentType);
+                    }
+                } catch (Exception e) {
                     e.printStackTrace();
-                    Log.e(LOG_TAG, e.getMessage());
+                } finally {
+                    urlConnection.disconnect();
                 }
             }
         }